requirements.txt
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
psycopg2
pytest==4.0.1
flask-paginate==0.5.2
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.17.0
coverage==4.5.3

